REFACTOR: 컬럼 내 Cell 3줄 이상 분리 관리
