+++ src/tables/OasisBotTransactionCompactColumns.tsx
@@ -1,9 +1,8 @@
 import { GridColDef, GridRenderCellParams, GridValidRowModel } from "@mui/x-data-grid";
 
-import ExchangeIcon from "@/components/Icon/ExchangeIcon";
-import ProfitChip from "@/components/chip/ProfitChip";
 import SellBuyChip from "@/components/chip/SellBuyChip";
-import { exchangeToKorean } from "@/libs/string";
+import TransactionColumnsMarketCell from "@/tables/row-cells/MarketCell";
+import TransactionColumnsProfitCell from "@/tables/row-cells/TransactionColumnsProfitCell";
 
 const OasisBotTransactionCompactColumns: GridColDef[] = [
   {
@@ -12,10 +11,7 @@ const OasisBotTransactionCompactColumns: GridColDef[] = [
     flex: 1,
     headerClassName: "text-slate-500",
     renderCell: (params: GridRenderCellParams<GridValidRowModel, ExchangeType>) => (
-      <div className="flex w-full justify-between whitespace-normal">
-        {exchangeToKorean(params.value)}
-        <ExchangeIcon exchange={params.value} size={20} />
-      </div>
+      <TransactionColumnsMarketCell {...params} />
     ),
   },
   {
@@ -33,17 +29,7 @@ const OasisBotTransactionCompactColumns: GridColDef[] = [
     flex: 1,
     headerClassName: "text-slate-500",
     renderCell: (params: GridRenderCellParams<GridValidRowModel, BotTransactionProfit>) => (
-      <div>
-        <div
-          className={`whitespace-normal ${params.value?.profitLoss && params.value.profitLoss > 0 ? "text-sub-3" : ""}`}
-        >
-          {params.value?.profitLoss ? "￦" : ""}
-          <span className="text-ellipsis font-bold">
-            {params.value?.profitLoss.toLocaleString("ko-kr") ?? "-"}
-          </span>
-        </div>
-        {params.value?.profitLossRate && <ProfitChip rate={params.value?.profitLossRate} />}
-      </div>
+      <TransactionColumnsProfitCell {...params} />
     ),
   },
   {
