+++ src/tables/row-cells/TransactionColumnsQuantityCell.tsx
@@ -1,0 +1,17 @@
+import { GridRenderCellParams, GridValidRowModel } from "@mui/x-data-grid";
+
+function TransactionColumnsQuantityCell(
+  params: GridRenderCellParams<GridValidRowModel, BotTransactionQuantity>,
+) {
+  return (
+    <div>
+      <div className="text-[16px] font-bold leading-[18px]">
+        ￦{params.value?.totalPrice.toLocaleString("ko-kr")}
+      </div>
+
+      <div className="mt-2 text-[14px] font-bold leading-[16px]">{params.value?.volume}</div>
+    </div>
+  );
+}
+
+export default TransactionColumnsQuantityCell;
